+++ v3/changelog.md
@@ -8,6 +8,9 @@
 
 ### Changed
 
+- Empty filter expressions are consistent between old and OpenDD execution
+  pipelines
+
 ### Fixed
 
 ## [v2025.03.10]